Undo config to disable entire Fabrication module
When disabled it causes creative menu to crash as it registers an empty Creative tab with a null item throwing NPEs.

As I'm unable to prevent it from generating tabs with the way it has been implemented here using Scala it is better to disable that feature.

Related to my previous PR: https://github.com/GTNewHorizons/ProjectRed/pull/39